logic for aol search for multiple groups of data

+++ src/main.py
@@ -67,8 +67,8 @@ def run(data_list, domain_list, output, settings):
                 p.starmap(partial(msmarco.to_search, qids=query_originals['qid'].values.tolist(), ranker=settings['ranker'], topk=100, batch=None), query_changes)
 
             # we need to add the original queries as well
-            # if not isfile(join(t5_output,f'original.{settings["ranker"]}')):
-            msmarco.to_search_df(pd.DataFrame(query_originals['query']), f'{t5_output}/original.{settings["ranker"]}', query_originals['qid'].values.tolist(), settings['ranker'], topk=100, batch=None)
+            if not isfile(join(t5_output, f'original.{settings["ranker"]}')):
+                msmarco.to_search_df(pd.DataFrame(query_originals['query']), f'{t5_output}/original.{settings["ranker"]}', query_originals['qid'].values.tolist(), settings['ranker'], topk=100, batch=None)
 
 
         if 'eval' in settings['cmd']:
@@ -105,7 +105,6 @@ def run(data_list, domain_list, output, settings):
         if not os.path.isdir(os.path.join(prep_index, 'indexes', index_item)): os.makedirs(os.path.join(prep_index, 'indexes', index_item))
         cat = True if 'docs' in {in_type, out_type} else False
         from dal import aol
-
         # AOL requires us to construct the Index, Qrels and Queries file from IR_dataset
         # create queries and qrels file
         aol.initiate_queries_qrels(prep_index)
@@ -114,13 +113,31 @@ def run(data_list, domain_list, output, settings):
         aol.create_json_collection(prep_index, index_item)
         create_index('aol', index_item)
         #to pair function
+        if not os.path.isfile(f'{prep_output}/queries.qrels.doc{"s" if cat else ""}.ctx.{index_item}.train.tsv'):
+            query_qrel_doc = aol.to_pair(prep_index, f'{prep_output}/queries.qrels.doc{"s" if cat else ""}.ctx.{index_item}.train.tsv', index_item,
+                                             cat=cat)
+            query_qrel_doc.to_csv(tsv_path['train'], sep='\t', encoding='utf-8', columns=[in_type, out_type],
+                                  index=False, header=False)
+            query_qrel_doc.to_csv(tsv_path['test'], sep='\t', encoding='utf-8', columns=[in_type, out_type],
+                                  index=False, header=False)
+        t5_model = settings['t5model']  # {"small", "base", "large", "3B", "11B"} cross {"local", "gc"}
+        t5_output = f'../output/{os.path.split(datapath)[-1]}/t5.{t5_model}.{index_item}.{in_type}.{out_type}'
+        if not os.path.isdir(t5_output): os.makedirs(t5_output)
+        if 'search' in settings['cmd']:
+            query_originals = pd.read_csv(f'{prep_output}/queries.qrels.doc{"s" if cat else ""}.ctx.{index_item}.train.tsv', sep='\t', usecols=['qid', 'query'], dtype={'qid': str})
+            query_changes = [(f'{t5_output}/{f}', f'{t5_output}/{f}.{settings["ranker"]}') for f in listdir(t5_output)
+                             if isfile(join(t5_output, f)) and f.startswith('pred.') and settings[
+                                 'ranker'] not in f and f'{f}.{settings["ranker"]}' not in listdir(t5_output)]
 
-        query_qrel_doc = aol.to_pair(prep_index, f'{prep_output}/queries.qrels.doc{"s" if cat else ""}.ctx.{index_item}.train.tsv', index_item,
-                                         cat=cat)
-        query_qrel_doc.to_csv(tsv_path['train'], sep='\t', encoding='utf-8', columns=[in_type, out_type],
-                              index=False, header=False)
-        query_qrel_doc.to_csv(tsv_path['test'], sep='\t', encoding='utf-8', columns=[in_type, out_type],
-                              index=False, header=False)
+            # for (i, o) in query_changes: msmarco.to_search(i, o, query_originals['qid'].values.tolist(), settings['ranker'], topk=100, batch=None)
+            # batch search: for (i, o) in query_changes: msmarco.to_search(i, o, query_originals['qid'].values.tolist(), settings['ranker'], topk=100, batch=2)
+            # parallel on each file
+            with multiprocessing.Pool(multiprocessing.cpu_count()) as p:
+                p.starmap(partial(aol.to_search, qids=query_originals['qid'].values.tolist(), index_item=index_item, ranker=settings['ranker'], topk=100, batch=None), query_changes)
+            # aol.to_search_df(pd.DataFrame(query_originals['query']), f'{t5_output}/original.{settings["ranker"]}', query_originals['qid'].values.tolist(),index_item, settings['ranker'], topk=100, batch=None)
+
+    if 'eval' in settings['cmd']:
+        from evl import trecw
     if ('yandex' in data_list): print('processing yandex...')
